added absolute path variable for tests folder to allow arbitrary binary placement; CTest now shows errors in case of failure; fixed all KiT-RT compiler warnings
Former-commit-id: 796f39b61cd65222581d1ce606e9523f70e31b4d

+++ code/src/toolboxes/cubic2dspline.cpp
@@ -18,11 +18,11 @@ double Cubic2DSpline::operator()( double x, double y ) {
     double points[4][4];
     for( int i = -1; i < 3; ++i ) {
         unsigned idx_y;
-        idx_y = yId + i < 0 ? 0 : yId + i;
+        idx_y = static_cast<int>( yId ) + i < 0 ? 0 : yId + i;
         idx_y = yId + i > _data.rows() - 1 ? _data.rows() - 1 : yId + i;
         for( int j = -1; j < 3; ++j ) {
             unsigned idx_x;
-            idx_x = xId + j < 0 ? 0 : xId + j;
+            idx_x = static_cast<int>( xId ) + j < 0 ? 0 : xId + j;
             idx_x = xId + j > _data.columns() - 1 ? _data.columns() - 1 : xId + j;
 
             points[i + 1][j + 1] = _data( idx_x, idx_y );
